+++ sdmetrics/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
-__version__ = '0.4.0'
+__version__ = '0.4.1.dev1'
 
 import pandas as pd
 
@@ -51,12 +51,14 @@ def compute_metrics(metrics, real_data, synthetic_data, metadata=None, **kwargs)
 
     scores = []
     for name, metric in metrics.items():
+        error = None
         try:
             raw_score = metric.compute(real_data, synthetic_data, **kwargs)
             normalized_score = metric.normalize(raw_score)
-        except Exception:
+        except Exception as err:
             raw_score = None
             normalized_score = None
+            error = str(err)
 
         scores.append({
             'metric': name,
@@ -66,6 +68,7 @@ def compute_metrics(metrics, real_data, synthetic_data, metadata=None, **kwargs)
             'min_value': metric.min_value,
             'max_value': metric.max_value,
             'goal': metric.goal.name,
+            'error': error,
         })
 
     return pd.DataFrame(scores)