+++ sdmetrics/timeseries/detection.py
@@ -77,8 +77,8 @@ class TimeSeriesDetectionMetric(TimeSeriesMetric):
             real_data, synthetic_data, metadata, entity_columns)
 
         transformer = rdt.HyperTransformer(default_data_type_transformers={
-            'O': 'one_hot_encoding',
-            'M': rdt.transformers.DatetimeTransformer(strip_constant=True),
+            'categorical': rdt.transformers.OneHotEncodingTransformer(error_on_unknown=False),
+            'datetime': rdt.transformers.DatetimeTransformer(strip_constant=True),
         })
         transformer.fit(real_data.drop(entity_columns, axis=1))